Add a CodeQueue to support predictable OAuth code responses
